splits extensions
